Backend/requirements.txt
elevenlabs
python-dotenv
python-multipart
fastapi
uvicorn[standard]
requests
supabase
boto3